Feature/studyroom frontend

See merge request s09-webmobile1-sub2/S09P12A206!87

+++ backend/sswm/src/main/java/com/ground/sswm/chat/domain/Chat.java
@@ -45,11 +45,11 @@ public class Chat {
     private String content;
 
     @Column(updatable = false)
-    private int createdAt;
+    private long createdAt;
 
 
     @Builder
-    public Chat(Studyroom studyroom, User user, String content, int createdAt) {
+    public Chat(Studyroom studyroom, User user, String content, long createdAt) {
         this.studyroom = studyroom;
         this.user = user;
         this.content = content;
